pyproject.toml
[project]
name = "ai-core"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20250625",
]
